Fix double listings when a package failed to build
Only run refresh_pkglist when a package has been built and installed
succesfully. When a package build failed (or was cancelled)
refresh_pkglist was called twice, once directly and once through
on_menuitem_unmarkall_activate, resulting in double listings of
packages. This only showed when a string was entered in the package
search box.