Fix dependency handling
Run slapt-src with the -n (--no-deps) option always. Since dependencies
have been calculated and have been put in the queue earlier, that means
that they have been processed already. No need to ask slapt-src to
calculate dependencies again.

Ignore %README% dependencies. slapt-src would fail, trying to find a
SlackBuild named "%README%". SBo folks put that in the REQUIRES field of
the .info file, when they want the user to actually read the README file
before running the SlackBuild. Of course there is no SlackBuild named
"%README%" and so slapt-src (and thus sourcery) would fail in this case.
We're just ignoring it. Users are advised to read the README file in all
cases anyway (and it's easy to do that from within sourcery).